Add main script and setup files.
This is the first version of the main script included with a setup script that can be ran for a slightly more user friendly way for someone to update their variables to the ones necessary for the bot to run on their server.
The context of the bot is to determine if users have a specific rule. If they do not have this role and join a Voice Channel they will be alerted that they do not have the role and directed to accept the rules.